+++ packages/xeus/src/index.ts
@@ -1,6 +1,4 @@
 export * from './coincident.worker';
 export * from './comlink.worker';
-export * from './interfaces';
-export * from './web_worker_kernel';
-export * from './worker.base';
+export * from './web.worker.kernel';
 export * from './worker';
